Move parse functionality to driver module
