+++ src/lib.rs
@@ -38,6 +38,7 @@
 extern crate phf;
 extern crate time;
 
+
 pub use html5ever_atoms::{Prefix, Namespace, LocalName, QualName};
 
 /// Re-export the tendril crate so that users don’t need to depend on it.
@@ -63,61 +64,8 @@ pub mod tokenizer;
 pub mod tree_builder;
 /// A simple reference-counted that serves as a default tree structure
 pub mod rcdom;
+/// Entrance to XML5 ever ParseResult
+pub mod driver;
 
-use tokenizer::{XmlTokenizerOpts, XmlTokenizer, TokenSink};
-use tree_builder::{TreeSink, XmlTreeBuilder};
-
-
-/// Parse and send results to a `TreeSink`.
-///
-/// ## Example
-///
-/// ```ignore
-/// let mut sink = MySink;
-/// parse_to(&mut sink, iter::once(my_str), Default::default());
-/// ```
-pub fn parse_to<
-        Sink:TreeSink,
-        It: IntoIterator<Item=tendril::StrTendril>
-    >(
-        sink: Sink,
-        input: It,
-        opts: XmlTokenizerOpts) -> Sink {
-
-    let tb = XmlTreeBuilder::new(sink);
-    let mut tok = XmlTokenizer::new(tb, opts);
-    for s in input {
-        tok.feed(s);
-    }
-    tok.end();
-    tok.unwrap().unwrap()
-}
-
-
-/// Parse into a type which implements `ParseResult`.
-///
-/// ## Example
-///
-/// ```ignore
-/// let dom: RcDom = parse(iter::once(my_str), Default::default());
-/// ```
-pub fn parse<Output, It>(input: It, opts: XmlTokenizerOpts) -> Output
-    where Output: ParseResult,
-          It: IntoIterator<Item=tendril::StrTendril>,
-{
-    let sink = parse_to(Default::default(), input, opts);
-    ParseResult::get_result(sink)
-}
-
-/// Results which can be extracted from a `TreeSink`.
-///
-/// Implement this for your parse tree data type so that it
-/// can be returned by `parse()`.
-pub trait ParseResult {
-    /// Type of consumer of tree modifications.
-    /// It also extends `Default` for convenience.
-    type Sink: TreeSink + Default;
-    /// Returns parsed tree data type
-    fn get_result(sink: Self::Sink) -> Self;
-}
-
+/// Re-export the encoding crate.
+pub use tendril::encoding;
